refactor: split non-step code from steps/tmux.rs to tmux.rs
Move Tmux struct, run_in_tmux, and run_command to src/tmux.rs,
keeping only run_tpm (the step) in src/steps/tmux.rs.

Closes #1755
